fix: split default tab logic (#38414)
## Description
Split the logic for the plugin action tabs. Now the files in CE and EE
repo for the `ce` folder are the same. Redefine the logic in EE hooks
only.

EE PR — https://github.com/appsmithorg/appsmith-ee/pull/5840

## Automation

/ok-to-test tags="@tag.All"

### :mag: Cypress test results
