+++ app/client/src/ee/PluginActionEditor/components/PluginActionResponse/hooks/useDefaultTab.tsx
@@ -1,0 +1,5 @@
+import { useDefaultTab as CE_useDefaultTab } from "ce/PluginActionEditor/components/PluginActionResponse/hooks/useDefaultTab";
+
+export function useDefaultTab() {
+  return CE_useDefaultTab();
+}
